TASK#122158 【后台】【电子处方】【广东医科大学附属医院】药品配送，药店取药第三方跳转链接地址修改优化

+++ ehealth-recipe-core/src/main/java/recipe/givemode/business/AbstractGiveModeService.java
@@ -160,7 +160,7 @@ public abstract class AbstractGiveModeService implements IGiveModeBase {
             return;
         }
         DrugsEnterprise drugsEnterprise = drugsEnterpriseDAO.getById(enterpriseId);
-        if (null == drugsEnterprise.getOrderType() || new Integer(1).equals(drugsEnterprise.getOrderType())) {
+        if (null == drugsEnterprise.getOrderType() || new Integer(0).equals(drugsEnterprise.getOrderType())) {
             return;
         }
         if (RecipeStatusEnum.RECIPE_STATUS_WAIT_SEND.getType().equals(recipe.getStatus())
